db blacklist fixes

+++ src/main/java/com/joy/NotificationService/controllers/MainController.java
@@ -11,8 +11,11 @@ import com.joy.NotificationService.services.MessageService;
 import com.joy.NotificationService.shared.dto.BlackListDto;
 import com.joy.NotificationService.shared.dto.EsDto;
 import com.joy.NotificationService.shared.dto.MessageDto;
+import com.joy.NotificationService.util.MessageStatus;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -30,41 +33,54 @@ public class MainController {
     ElasticSearchService elasticSearchService;
 
     @PostMapping("v1/sms/send")
-    public MessageModelResponse send(@RequestBody Message msg) {
+    public ResponseEntity<Object> send(@RequestBody Message msg) {
+
+        if(msg.getMessage().trim().isEmpty()){
+            return new ResponseEntity<>(new MessageModelResponse(new ErrorResponse("INVALID_REQUEST","Message cannot be empty")), HttpStatus.BAD_REQUEST);
+        }
+        if(msg.getPhone_number().trim().isEmpty()){
+            return new ResponseEntity<>(new MessageModelResponse(new ErrorResponse("INVALID_REQUEST","phone_number is mandatory")), HttpStatus.BAD_REQUEST);
+        }
         MessageDto msgDto = messageService.storeRequest(msg);
-        if (msgDto != null) {
+        if (msgDto!=null) {
             DataResponse data = new DataResponse();
             data.setRequest_id(msgDto.getId());
             data.setComments("Successfully Sent");
             MessageModelResponse returnValue = new MessageModelResponse(data);
-            return returnValue;
+            return new ResponseEntity<>(returnValue,HttpStatus.OK);
         }
-        ErrorResponse error = new ErrorResponse();
-        error.setCode("Error...");
-        error.setMessage("Server down...");
-        MessageModelResponse returnValue = new MessageModelResponse(error);
-        return returnValue;
+//        ErrorResponse error = new ErrorResponse();
+//        error.setCode("Error...");
+//        error.setMessage("Server down...");
+//        MessageModelResponse returnValue = new MessageModelResponse(error);
+        return new ResponseEntity<>(new MessageModelResponse(new ErrorResponse("Bad Service","Server Down")),HttpStatus.INTERNAL_SERVER_ERROR);
     }
 
     @PostMapping("v1/blacklist")
-    public BlackListAddResponse addNumbers(@RequestBody BlackListNumbers blackListNumbers){
+    public ResponseEntity<Object> addNumbers(@RequestBody BlackListNumbers blackListNumbers){
+        if(blackListNumbers.getPhone_numbers().isEmpty()){
+            return new ResponseEntity<>(new ErrorResponse("INVALID_REQUEST","No number to add"),HttpStatus.BAD_REQUEST);
+        }
         blackListService.saveNumber(blackListNumbers);
         BlackListAddResponse returnValue=new BlackListAddResponse();
         returnValue.setData("Successfully Blacklisted");
-        return returnValue;
+
+        return new ResponseEntity<>(returnValue,HttpStatus.OK);
     }
 
     @DeleteMapping("v1/blacklist")
-    public BlackListAddResponse deleteNumber(@RequestBody BlackListNumber blackListNumber){
+    public ResponseEntity<Object> deleteNumber(@RequestBody BlackListNumber blackListNumber){
+        if(blackListNumber.getPhone_number().isEmpty())
+            return new ResponseEntity<>(new ErrorResponse("INVALID_REQUEST","phone_number cannot be empty"),HttpStatus.BAD_REQUEST);
         boolean temp=blackListService.deleteNumber(blackListNumber.getPhone_number());
         BlackListAddResponse returnValue=new BlackListAddResponse();
         if(temp){
             returnValue.setData("Successfully Deleted");
+            return new ResponseEntity<>(returnValue,HttpStatus.OK);
         }
-        else{
-            returnValue.setData("Error: Number not Blacklisted...");
-        }
-        return returnValue;
+
+        returnValue.setData("Error: Number not Blacklisted...");
+        return new ResponseEntity<>(returnValue,HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping("v1/blacklist")
@@ -73,12 +89,19 @@ public class MainController {
         return returnValue;
     }
 
-    @GetMapping("/v1/sms/{id}")
-    public MessageDetailResponse getMessageDetail(@PathVariable Integer id){
+    @GetMapping("v1/sms/{id}")
+    public ResponseEntity<Object> getMessageDetail(@PathVariable Integer id){
+
         MessageDto messageDto=messageService.getMessageDetail(id);
-        MessageDetailResponse returnValue=new MessageDetailResponse();
-        returnValue.setData(messageDto);
-        return returnValue;
+        System.out.println(messageDto);
+
+        if(messageDto!=null){
+            MessageDetailResponse returnValue=new MessageDetailResponse();
+            returnValue.setData(messageDto);
+            return new ResponseEntity<>(returnValue,HttpStatus.OK);
+        }
+        return new ResponseEntity<>(new MessageModelResponse(
+                new ErrorResponse("INVALID_REQUEST","request_id not found")),HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping ("/v1/message/{text}")
